requirements.txt
aiosqlite==0.19.0 ; python_version >= "3.9" and python_version < "3.11"
annotated-types==0.5.0 ; python_version >= "3.9" and python_version < "3.11"
anyio==3.7.1 ; python_version >= "3.9" and python_version < "3.11"
asn1crypto==1.5.1 ; python_version >= "3.9" and python_version < "3.11"
bcrypt==4.0.1 ; python_version >= "3.9" and python_version < "3.11"
certifi==2023.7.22 ; python_version >= "3.9" and python_version < "3.11"
cffi==1.15.1 ; python_version >= "3.9" and python_version < "3.11"
charset-normalizer==3.2.0 ; python_version >= "3.9" and python_version < "3.11"
click==8.1.7 ; python_version >= "3.9" and python_version < "3.11"
cloudpickle==2.0.0 ; python_version >= "3.9" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "3.11" and (sys_platform == "win32" or platform_system == "Windows")
cryptography==41.0.4 ; python_version >= "3.9" and python_version < "3.11"
databases[sqlite]==0.8.0 ; python_version >= "3.9" and python_version < "3.11"
dnspython==2.4.2 ; python_version >= "3.9" and python_version < "3.11"
email-validator==2.0.0.post2 ; python_version >= "3.9" and python_version < "3.11"
exceptiongroup==1.1.3 ; python_version >= "3.9" and python_version < "3.11"
fastapi-users-db-sqlalchemy==4.0.5 ; python_version >= "3.9" and python_version < "3.11"
fastapi-users==11.0.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi-users[sqlalchemy]==11.0.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.103.1 ; python_version >= "3.9" and python_version < "3.11"
filelock==3.12.4 ; python_version >= "3.9" and python_version < "3.11"
greenlet==2.0.2 ; python_version >= "3.9" and python_version < "3.11"
gunicorn==21.2.0 ; python_version >= "3.9" and python_version < "3.11"
h11==0.14.0 ; python_version >= "3.9" and python_version < "3.11"
idna==3.4 ; python_version >= "3.9" and python_version < "3.11"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "3.11"
joblib==1.3.2 ; python_version >= "3.9" and python_version < "3.11"
loguru==0.7.2 ; python_version >= "3.9" and python_version < "3.11"
makefun==1.15.1 ; python_version >= "3.9" and python_version < "3.11"
markupsafe==2.1.3 ; python_version >= "3.9" and python_version < "3.11"
numpy==1.26.0 ; python_version >= "3.9" and python_version < "3.11"
oscrypto==1.3.0 ; python_version >= "3.9" and python_version < "3.11"
packaging==23.1 ; python_version >= "3.9" and python_version < "3.11"
pandas==2.0.3 ; python_version >= "3.9" and python_version < "3.11"
passlib[bcrypt]==1.7.4 ; python_version >= "3.9" and python_version < "3.11"
platformdirs==3.8.1 ; python_version >= "3.9" and python_version < "3.11"
plotly==5.17.0 ; python_version >= "3.9" and python_version < "3.11"
pyarrow==10.0.1 ; python_version >= "3.9" and python_version < "3.11"
pycparser==2.21 ; python_version >= "3.9" and python_version < "3.11"
pycryptodomex==3.19.0 ; python_version >= "3.9" and python_version < "3.11"
pydantic-core==2.10.1 ; python_version >= "3.9" and python_version < "3.11"
pydantic==2.4.1 ; python_version >= "3.9" and python_version < "3.11"
pyjwt==2.6.0 ; python_version >= "3.9" and python_version < "3.11"
pyjwt[crypto]==2.6.0 ; python_version >= "3.9" and python_version < "3.11"
pyopenssl==23.2.0 ; python_version >= "3.9" and python_version < "3.11"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "3.11"
python-decouple==3.8 ; python_version >= "3.9" and python_version < "3.11"
python-multipart==0.0.6 ; python_version >= "3.9" and python_version < "3.11"
pytz==2023.3.post1 ; python_version >= "3.9" and python_version < "3.11"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "3.11"
requests==2.31.0 ; python_version >= "3.9" and python_version < "3.11"
scikit-learn==1.3.1 ; python_version >= "3.9" and python_version < "3.11"
scipy==1.11.2 ; python_version >= "3.9" and python_version < "3.11"
setuptools==68.2.2 ; python_version >= "3.9" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.9" and python_version < "3.11"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "3.11"
snowflake-connector-python==3.2.0 ; python_version >= "3.9" and python_version < "3.11"
snowflake-connector-python[pandas]==3.2.0 ; python_version >= "3.9" and python_version < "3.11"
snowflake-snowpark-python[pandas]==1.8.0 ; python_version >= "3.9" and python_version < "3.11"
snowflake-sqlalchemy==1.5.0 ; python_version >= "3.9" and python_version < "3.11"
sortedcontainers==2.4.0 ; python_version >= "3.9" and python_version < "3.11"
sqlalchemy==1.4.49 ; python_version >= "3.9" and python_version < "3.11"
sqlalchemy[asyncio]==1.4.49 ; python_version >= "3.9" and python_version < "3.11"
starlette==0.27.0 ; python_version >= "3.9" and python_version < "3.11"
tenacity==8.2.3 ; python_version >= "3.9" and python_version < "3.11"
threadpoolctl==3.2.0 ; python_version >= "3.9" and python_version < "3.11"
tomlkit==0.12.1 ; python_version >= "3.9" and python_version < "3.11"
tqdm==4.66.1 ; python_version >= "3.9" and python_version < "3.11"
typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "3.11"
urllib3==1.26.16 ; python_version >= "3.9" and python_version < "3.11"
uvicorn==0.23.2 ; python_version >= "3.9" and python_version < "3.11"
wheel==0.41.2 ; python_version >= "3.9" and python_version < "3.11"
win32-setctime==1.1.0 ; python_version >= "3.9" and python_version < "3.11" and sys_platform == "win32"
